Refactored validator to share code between the config and the template validations
